asset/templates: move openshift manifest templates from manifests/tectonic to manifests/openshift
Changes for https://jira.coreos.com/browse/CORS-878
