+++ pkg/asset/templates/content/openshift/binding-discovery.go
@@ -1,4 +1,4 @@
-package tectonic
+package openshift
 
 import (
 	"os"
@@ -9,31 +9,31 @@ import (
 )
 
 const (
-	cloudCredsSecretFileName = "cloud-creds-secret.yaml.template"
+	bindingDiscoveryFileName = "binding-discovery.yaml"
 )
 
-var _ asset.WritableAsset = (*CloudCredsSecret)(nil)
+var _ asset.WritableAsset = (*BindingDiscovery)(nil)
 
-// CloudCredsSecret is the constant to represent contents of corresponding yaml file
-type CloudCredsSecret struct {
+// BindingDiscovery  is the variable/constant representing the contents of the respective file
+type BindingDiscovery struct {
 	fileName string
 	FileList []*asset.File
 }
 
 // Dependencies returns all of the dependencies directly needed by the asset
-func (t *CloudCredsSecret) Dependencies() []asset.Asset {
+func (t *BindingDiscovery) Dependencies() []asset.Asset {
 	return []asset.Asset{}
 }
 
 // Name returns the human-friendly name of the asset.
-func (t *CloudCredsSecret) Name() string {
-	return "CloudCredsSecret"
+func (t *BindingDiscovery) Name() string {
+	return "BindingDiscovery"
 }
 
 // Generate generates the actual files by this asset
-func (t *CloudCredsSecret) Generate(parents asset.Parents) error {
-	t.fileName = cloudCredsSecretFileName
-	data, err := content.GetTectonicTemplate(t.fileName)
+func (t *BindingDiscovery) Generate(parents asset.Parents) error {
+	t.fileName = bindingDiscoveryFileName
+	data, err := content.GetOpenshiftTemplate(t.fileName)
 	if err != nil {
 		return err
 	}
@@ -47,13 +47,13 @@ func (t *CloudCredsSecret) Generate(parents asset.Parents) error {
 }
 
 // Files returns the files generated by the asset.
-func (t *CloudCredsSecret) Files() []*asset.File {
+func (t *BindingDiscovery) Files() []*asset.File {
 	return t.FileList
 }
 
 // Load returns the asset from disk.
-func (t *CloudCredsSecret) Load(f asset.FileFetcher) (bool, error) {
-	file, err := f.FetchByName(filepath.Join(content.TemplateDir, cloudCredsSecretFileName))
+func (t *BindingDiscovery) Load(f asset.FileFetcher) (bool, error) {
+	file, err := f.FetchByName(filepath.Join(content.TemplateDir, bindingDiscoveryFileName))
 	if err != nil {
 		if os.IsNotExist(err) {
 			return false, nil
